refactor(test): consolidate utility tests and add cursor unit tests
Merge tests/unit/test_uri_template.py into tests/unit/test_utils.py to reduce file sprawl. Add direct unit test coverage for litestar_mcp/_cursor.py in tests/unit/test_utils.py. Delete the redundant test_uri_template.py.

TAG=agy
CONV=4db20f37-0e22-4128-983c-20e382b47363
